Cache logic now uses ADAL::UserIdentifier.
Previously, the cache just compared the field `user_id` of ADAL::TokenResponse as a string. Now it is its own class with its own sense of equality. Furthermore, AuthenticationContext#acquire_token_for_user now fully supports UserIdentifier. However, if there is no matching token in the cache, it throws a UserCredentialError with a descriptive error message.